+++ cmake/AISInstall.cmake
@@ -8,7 +8,7 @@
 include(ROCMInstallTargets)
 include(ROCMCreatePackage)
 
-# Install the package
+# Install the library
 rocm_install(TARGETS hipfile)
 
 # Install the headers
@@ -17,8 +17,37 @@ rocm_install(
     DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
 
-# Install utilities
-install(PROGRAMS util/aischeck DESTINATION bin)
+# Install AIS tools
+install(PROGRAMS tools/ais-check/ais-check DESTINATION bin)
+
+# Install example code
+# Since the input DIRECTORY is `examples` don't include it in
+# the DESTINATION path or you'll get `examples/examples/*` output
+if(AIS_INSTALL_EXAMPLES)
+    install(
+        DIRECTORY ${CMAKE_SOURCE_DIR}/examples
+        DESTINATION share/doc/${CMAKE_PROJECT_NAME}
+        FILES_MATCHING
+            PATTERN "*.cpp"
+            PATTERN "*.h"
+            PATTERN "README.md"
+    )
+
+    # Create the installed CMakeLists.txt files from the templates.
+    # The CMakeLists.txt files in the examples tree use our build
+    # flags and infrastructure so we can ensure they are well-vetted
+    # and these can't be installed.
+
+    # aiscp
+    configure_file(
+        "${CMAKE_CURRENT_SOURCE_DIR}/examples/aiscp/CMakeLists.install.in"
+        "examples/aiscp/CMakeLists.txt"
+    )
+    install(FILES
+        "${CMAKE_CURRENT_BINARY_DIR}/examples/aiscp/CMakeLists.txt"
+        DESTINATION "share/doc/${CMAKE_PROJECT_NAME}/examples/aiscp/"
+    )
+endif()
 
 # When we have RELEASE/DEV builds set up, we can split
 # where these dependencies are added.
@@ -46,23 +75,26 @@ if(CMAKE_HIP_PLATFORM STREQUAL "amd")
     rocm_package_add_rpm_dependencies(DEPENDS libmount-devel)
 endif()
 
-# Nvidia Runtime Dependencies
-if(CMAKE_HIP_PLATFORM STREQUAL "nvidida")
+# NVIDIA Runtime Dependencies
+if(CMAKE_HIP_PLATFORM STREQUAL "nvidia")
     rocm_package_add_dependencies(DEPENDS libcufile)
 endif()
 
-# Nvidia Development Dependencies
+# NVIDIA Development Dependencies
 if(CMAKE_HIP_PLATFORM STREQUAL "nvidia")
     rocm_package_add_deb_dependencies(DEPENDS libcufile-dev)
     rocm_package_add_rpm_dependencies(DEPENDS libcufile-devel)
 endif()
 
-# Export the targets
-set(target_list ${target_list} roc::hipfile)
+# Export the hipfile target
+set(target_list hip::hipfile)
 
+# Kitware recommends making the namespace match the package name
+# to work with the Common Package Specification, but since the
+# rest of ROCm doesn't do that, we'll stick with hip::
 rocm_export_targets(
     TARGETS ${target_list}
-    NAMESPACE roc::
+    NAMESPACE hip::
 )
 
 # CPack license setup