+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## UNRELEASED - hipFile 0.2.0
 ### Added
+* The examples can be installed to `share/doc/examples/*` by setting `AIS_INSTALL_EXAMPLES` (formerly `AIS_BUILD_EXAMPLES`) to ON
 
 ### Changed
 * Renamed `hipFileOpStatusError()` to `hipFileGetOpErrorString()`
 * The `hipfile-doc` CMake target no longer exists (just use `doc`)
 * The `doc` CMake target only exists if the `AIS_BUILD_DOCS` option is enabled
 * `hipFileRead()`/`hipFileWrite()` will transfer at most 0x7ffff000 (2,147,479,552) bytes returning the number of bytes actually transferred
+* The CMake namespace was changed from `roc::` to `hip::`
+* `AIS_BUILD_EXAMPLES` has been renamed to `AIS_INSTALL_EXAMPLES`
 
 ### Removed
 * The rocFile library has been completely removed and the code is now a part of hipFile.